Revamped Tree, Refined Grab Distance & Lowered Mouse Sensitivity
The tree was created from scratch to take advantage of a better design idea I had. 
The mouse sensitivity was lowered dramatically so as to compensate for it being greatly increased when run in Web GL.
Hopefully the distance from a Thought a player needs to be in order to grab it is increased enough so that they aren't so awkward to grab anymore.